Split apart DataSS into DataPP & DataSS
